fixed conn.c

+++ trunk/devel/libsbase/src/conn.c
@@ -568,7 +568,7 @@ int conn_write_handler(CONN *conn)
 #ifdef HAVE_SSL
                 if(conn->ssl) ERR_print_errors_fp(stdout);
 #endif
-                FATAL_LOGGER(conn->logger, "Sending chunk[%d-%d] data to %s:%d on %s:%d via %d failed, %s", CKN(cp), CK_LEFT(cp), conn->remote_ip, conn->remote_port, conn->local_ip, conn->local_port, conn->fd, strerror(errno));
+                FATAL_LOGGER(conn->logger, "Sending chunk[%d-%lld] data to %s:%d on %s:%d via %d failed, %s", CKN(cp), (long long)(CK_LEFT(cp)), conn->remote_ip, conn->remote_port, conn->local_ip, conn->local_port, conn->fd, strerror(errno));
                 /* Terminate connection */
                 CONN_TERMINATE(conn, D_STATE_CLOSE);
             }
